icnet/main/requirements.txt
scipy
matplotlib 
numpy
tqdm
pillow

